200855
Defer loading the server certificates until the 2nd module load. Otherwise
any problems with said certificate(s) will not be displayed because
the VirtualHost logging hasn't been turned on yet.

Switch a slew of APLOG_INFO to APLOG_ERR.

And add a default LogLevel to nss.conf.in. This is not inherited from
httpd.conf.