Add some basic functional tests.
This tests in an in-tree Apache instance using the local libmodnss.so
shared library, so no pre-installation is necessary.

The tests use python-nose and a hacked python-requests library. It is
hacked so I can obtain the negotiated cipher and protocol as well as
pass a few other things into it.

Tests right now are limited to GET requests.

A new user certificate for 'beta' was added to gencert to do pass/fail
access control testing.

The basic process of the tests are:

- run setup.sh which sets up a new instance with createinstance.sh
  and does some variable substitution.
- nosetests -v

I picture multiple test "suites" of different configurations. Right now
there is only one. A template file is provided for each suite.

Tested only on Fedora 20 right now.