Add test suite for SNI
python for OpenSSL is in quite a sad state with several competing
mid-level implementations which provide different feature sets.

The httplib client provides access to the negotiated cipher and
protocol but not SNI (and it has lousy hostname checking).

The urllib3 client provides SNI and is generally better but doesn't
give any details on the connection.

So I'm using both. The original one is used for basic server testing
and the urllib3 one is used just for SNI testing.

Also:

- Indent the test configuration to make it more readable
- Add separate config file for SNI testing
- Add a CGI configuration and script to test CGI variables
- Change client cipher test to use AES256-SHA instead of RC4
- Add a commented-out valgrind option in start for future
  debuggers
- Change the VirtualServers to *:port and use ServerName
- Add per-VH document roots so SNI can be more easily tested