Rework SNI client reverse proxy
Add a note to the table to indicate that the handhake is complete
so we don't set the extension every time data is read or written.

Drop NSSHandshakeCallback() as it didn't do anything and is replaced
by the proxy callback.

Extend the checks around calling SetURL to match those in mod_ssl:
  - a hostname is available
  - not SSLv3
  - not an IP address