Always call SSL_ShutdownServerSessionIDCache() in ModuleKill
The guard of NSS_IsInitialized() was too strict because of the
way Apache loads and unloads modules. We need to clean up the
SessionIDCache thread locking when a SIGHUP is received otherwise
a crash will occur.

Note that this also eliminates a rather huge memory leak when
the server is reloaded with a SIGHUP.

https://bugzilla.redhat.com/show_bug.cgi?id=1277613
https://bugzilla.redhat.com/show_bug.cgi?id=1295976
https://fedorahosted.org/mod_nss/ticket/16