Don't ignore NSSProtocol when NSSFIPS is enabled
The value was always being set to TLS 1.0, 1.1 and 1.2, ignoring
the configuration value.

I suspect this is because this code dated to when only SSL2, 3 and
TLS 1.0 were supported so it only enabled TLS v1.0. When 1.1 and
1.2 were added it seemed natural to automatically enable those
as well. Natural but incorrect.

Based on patch by Matthew Harmsen <mharmsen@redhat.com>

RHBZ #1312052