Add basic cipher aliases to fix anti-patterns/typos
A couple of ciphers didn't follow the naming pattern and
used _sha256 instead of sha_256 and I mis-spelled camellia
in two of them.

Add a new element to the cipher table, alias. It can contain
a single entry. It could be extended to a CSV if needed but I
don't see the point in doing that work now.

https://fedorahosted.org/mod_nss/ticket/26